Changes in kd_analyze.py: dict key is now a tuple;  Also, added char_freq2.pdf into docs
In kd_analyze.py, I was using a string "34:32" as the key, but it's easier to do something
with if it's a tuple (34,32).  So that's what I'm doing now.  Also, added some functions
that make it easier to print the actual character to the screen based on the ascii code.

The "char_freq2.pdf" is a document I found online that shows the frequency of character
pairs in the english language.  We could pick the most frequent few to use in our
analysis, perhaps.